性能优化

+++ src/stealbomber/attack/password.java
@@ -33,7 +33,7 @@ public class password {
     private static void random() {
         final String str = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@#$%^&*";
         StringBuilder pass = new StringBuilder();
-        for (int i = 0; i < rp; i++) {
+        for (byte i = 0; i < rp; i++) {
             int randomInt = random.nextInt(str.length());
             pass.append(str.charAt(randomInt));
         }
@@ -43,7 +43,7 @@ public class password {
     private static void randomNoSpecial() {
         final String str = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
         StringBuilder pass = new StringBuilder();
-        for (int i = 0; i < rp; i++) {
+        for (byte i = 0; i < rp; i++) {
             int randomInt = random.nextInt(str.length());
             pass.append(str.charAt(randomInt));
         }
@@ -55,10 +55,10 @@ public class password {
         final String letter = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
         StringBuilder pass = new StringBuilder();
         int letterlong = random.nextInt(1) + 3;
-        for (int i = 0; i < letterlong; i++) {
+        for (byte i = 0; i < letterlong; i++) {
             pass.append(letter.charAt(random.nextInt(letter.length())));
         }
-        for (int i = 0; i < (rp - letterlong); i++) {
+        for (byte i = 0; i < (rp - letterlong); i++) {
             pass.append(num.charAt(random.nextInt(num.length())));
         }
         password = pass.toString();
@@ -69,10 +69,10 @@ public class password {
         final String letter = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
         StringBuilder pass = new StringBuilder();
         int letterlong = random.nextInt(1) + 3;
-        for (int i = 0; i < (rp - letterlong); i++) {
+        for (byte i = 0; i < (rp - letterlong); i++) {
             pass.append(num.charAt(random.nextInt(num.length())));
         }
-        for (int i = 0; i < letterlong; i++) {
+        for (byte i = 0; i < letterlong; i++) {
             pass.append(letter.charAt(random.nextInt(letter.length())));
         }
         password = pass.toString();
@@ -81,7 +81,7 @@ public class password {
     private static void num() {
         final String num = "0123456789";
         StringBuffer pass = new StringBuffer();
-        for (int i = 0; i < rp; i++) {
+        for (byte i = 0; i < rp; i++) {
             pass.append(num.charAt(random.nextInt(num.length())));
         }
         password = pass.toString();
